Ruslan Test (new line)

+++ src/main/java/students/ruslan_k/testArifmetic.java
@@ -4,7 +4,7 @@ public class testArifmetic {
     public static void main(String[] args) {
         int a = 10;
         int b = 4;
-        double result = a / b;
+        double result = (double)a / (double)b;
 
         System.out.println(result);
     }
